+++ src/native_client/mod.rs
@@ -69,6 +69,7 @@ fn connect<'py>(
             handshake_done: false,
             handshake_fut: None,
             expected_accept,
+            pending_pings: Default::default(),
             pending_recv: VecDeque::new(),
             backlog: VecDeque::new(),
             on_message,